- Updated Factory to implement the 'int' interface over 'double' interface with the updated ResourcePool. - Added a new constant to Numbers. - Updated SeedSmelter logic to provide cleaner numbers when processing raw ores. - Updated SeedSmelter logic so that it will only pull raw ores when there are enough of them (at this time, 10 units of ore).
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@51 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/Things/Structures/Factory.h
@@ -13,7 +13,7 @@ struct ProductionCost
 {
 	ProductionCost() : TurnsToBuild(0) {}
 
-	ProductionCost(int turns, double commonMetals, double commonMinerals, double rareMetals, double rareMinerals) : TurnsToBuild(turns)
+	ProductionCost(int turns, int commonMetals, int commonMinerals, int rareMetals, int rareMinerals) : TurnsToBuild(turns)
 	{
 		CostPerTurn.commonMetals(commonMetals);
 		CostPerTurn.commonMinerals(commonMinerals);
